requirements.txt
Django==1.11.1
django-filter==1.0.4
djangorestframework==3.6.3
psycopg2==2.7.1
pytz==2017.2
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
cffi==1.10.0
constantly==15.1.0
cryptography==1.8.2
cssselect==1.0.1
Django==1.11.1
django-filter==1.0.4
djangorestframework==3.6.3
idna==2.5
incremental==17.5.0
lxml==3.7.3
numpy==1.12.1
packaging==16.8
pandas==0.20.1
parsel==1.2.0
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==17.0.0
pyparsing==2.2.0
pypiwin32==220
python-dateutil==2.6.0
pytz==2017.2
queuelib==1.4.2
Scrapy==1.4.0
scrapy-djangoitem==1.1.1
service-identity==17.0.0
six==1.10.0
Twisted==17.1.0
w3lib==1.17.0
zope.interface==4.4.1
